+++ CMakeLists.txt
@@ -628,7 +628,7 @@ source_group("imgui" FILES ${imguiSrc})
 set(SRC_SHP)
 set(SRC_GUI)
 set(SRC_UTL)
-set(SRC_SKETCH)
+set(SRC_SKT)
 set(SRC_SCR)
 set(SRC_OTHER)
 foreach(f ${srcs})
@@ -639,8 +639,8 @@ foreach(f ${srcs})
     list(APPEND SRC_GUI ${f})
   elseif(_src_name MATCHES "^utl(_|\\.)")
     list(APPEND SRC_UTL ${f})
-  elseif(_src_name MATCHES "^sketch")
-    list(APPEND SRC_SKETCH ${f})
+  elseif(_src_name MATCHES "^skt")
+    list(APPEND SRC_SKT ${f})
   elseif(_src_name MATCHES "^scr")
     list(APPEND SRC_SCR ${f})
   else()
@@ -650,7 +650,7 @@ endforeach()
 source_group("src\\shp" FILES ${SRC_SHP})
 source_group("src\\gui" FILES ${SRC_GUI})
 source_group("src\\utl" FILES ${SRC_UTL})
-source_group("src\\sketch" FILES ${SRC_SKETCH})
+source_group("src\\skt" FILES ${SRC_SKT})
 source_group("src\\scr" FILES ${SRC_SCR})
 source_group("src" FILES ${SRC_OTHER})
 source_group("web" FILES ${WEB_FILES_REL})
