+++ src/doc_delta.cpp
@@ -2,7 +2,7 @@
 
 #include "gui_occt_view.h"
 #include "sketch.h"
-#include "sketch_json.h"
+#include "skt_json.h"
 #include "utl.h"
 
 Sketch_struct_delta::Sketch_struct_delta(Kind kind, nlohmann::json sketch_json, bool was_current,
