+++ src/gui_occt_view.cpp
@@ -59,8 +59,8 @@
 #include "shp_create.h"
 #include "shp_delta.h"
 #include "sketch.h"
-#include "sketch_ais.h"
-#include "sketch_json.h"
+#include "skt_ais.h"
+#include "skt_json.h"
 #include "utl_types.h"
 #include "utl.h"
 #include "utl_json.h"
