+++ src/gui_settings.cpp
@@ -11,7 +11,7 @@
 #include "gui_occt_view.h"
 #include "utl_settings.h"
 #include "sketch.h"
-#include "sketch_nodes.h"
+#include "skt_nodes.h"
 
 namespace
 {
