+++ src/shp_polar_dup.cpp
@@ -12,7 +12,7 @@
 #include "mode.h"
 #include "gui_occt_view.h"
 #include "sketch.h"
-#include "sketch_nodes.h"
+#include "skt_nodes.h"
 #include "shp_delta.h"
 
 Shp_polar_dup::Shp_polar_dup(Occt_view& view)
