+++ src/skt_ais.h
