+++ src/skt_edge.h
