+++ src/skt_nodes.cpp
@@ -1,4 +1,4 @@
-#include "sketch_nodes.h"
+#include "skt_nodes.h"
 
 #include <BRep_Builder.hxx>
 #include <BRepBuilderAPI_MakeEdge.hxx>
