+++ src/skt_nodes.h
