+++ src/skt_op_recorder.h
