+++ src/skt_operations.cpp
@@ -7,8 +7,8 @@
 
 #include "gui.h"
 #include "mode.h"
-#include "sketch_op_recorder.h"
-#include "sketch_edge.h"
+#include "skt_op_recorder.h"
+#include "skt_edge.h"
 #include "utl_geom.h"
 #include "utl_occt.h"
 #include "utl.h"
