+++ src/skt_tools.h
@@ -4,7 +4,7 @@
 #include <optional>
 #include <vector>
 
-#include "sketch_edge.h"
+#include "skt_edge.h"
 #include "utl_types.h"
 
 class Sketch;
