+++ src/skt_underlay.cpp
@@ -1,4 +1,4 @@
-#include "sketch_underlay.h"
+#include "skt_underlay.h"
 
 #include <AIS_InteractiveObject.hxx>
 
@@ -25,7 +25,7 @@
 #include "utl_geom.h"
 #include "utl_asset_store.h"
 #include "gui_occt_view.h"
-#include "sketch_nodes.h"
+#include "skt_nodes.h"
 #include "utl.h"
 
 using namespace glm;
