+++ tests/skt_topo_tests.cpp
@@ -1,4 +1,4 @@
-#include "sketch_test_fixture.h"
+#include "skt_test_fixture.h"
 
 #include <BRepGProp.hxx>
 #include <BRepTools.hxx>
@@ -11,8 +11,8 @@
 #include <numbers>
 #include <vector>
 
-#include "sketch_edge.h"
-#include "sketch_nodes.h"
+#include "skt_edge.h"
+#include "skt_nodes.h"
 #include "utl_geom.h"
 
 using namespace glm;
